Update customdashboard.md
The question "How is Advanced mode different from Query mode while writing queries?" was added twice. Modified the question to "If a query is written using the Advanced mode, will I be able to view the same query in Query Mode as well?" and also rewritten the answer for this question.